Create main.py
This is a tournament builder for a fried of mine. He has Loyalists, Traitors and a small group of Neutrals who can be either.

The idea is to have winners v winners, losers v losers - with the Neutrals filling the slots.

Ideally, they should not play on the same board. If they do, print a message before doing the wins so they can be shuffled.